Move create-account files to corresponding module Fix errors and imports
