+++ app/build.gradle.kts
@@ -87,5 +87,6 @@ dependencies {
     implementation(project(":shared:designsystem:resources"))
 
     implementation(project(":features:auth:login"))
+    implementation(project(":features:auth:create-account"))
     implementation(project(":features:settings"))
 }
