+++ features/auth/create-account/src/main/res/values/strings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="create_account">Welcome to Krypt. Let\'s get started!</string>
+    <string name="successfully_create_account">Account created successfully</string>
+    <string name="button_create_account">Create Account</string>
+
+    <string name="password_length_error">Password length must be greater than twelve characters</string>
+    <string name="account_length_error">Account name length must be greater than four characters</string>
+
+</resources>
